almost complete
added the settings and fullscreen buttons, tweaked the popups (save
asking and settings), and also gunna try the rolling release version of
the service worker :grinning:

Reason for having the fullscreen button on the main screen, is that at
least with Android, while in fullscreen, swiping from top or bottom
edges shows he android UI with home & back buttons which would stop the
fullscreen anyway, so best to teach toddler to not press it. That is a
VERY good lesson to learn; what NOT to touch!