nope... again!
that simply does not work... because of chrome-extensions by the seem of
it. Very annoying, because the cache version requires a full redownload
of everything in the cache regardless of whether a file has changed or
not.

Hopefully I'll find a better way than just creating loads of caches for
frequently and non-frequently updated files!